refactor: change embedding and base node
