+++ docarray/predefined_document/text.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from docarray.document import BaseDocument
-from docarray.typing.ndarray import Embedding, Tensor
+from docarray.typing.embedding import Embedding, Tensor
 
 
 class Text(BaseDocument):
