+++ docarray/typing/__init__.py
@@ -1,6 +1,5 @@
-from docarray.document.base_node import BaseNode
-
-from docarray.typing.ndarray import Embedding, Tensor
+from docarray.typing.embedding import Embedding
+from docarray.typing.tensor import Tensor
 from docarray.typing.url import ImageUrl
 
-__all__ = ['Tensor', 'Embedding', 'BaseNode', 'ImageUrl']
+__all__ = ['Tensor', 'Embedding', 'ImageUrl']
