+++ docarray/typing/embedding.py
@@ -1,0 +1,5 @@
+from docarray.typing.tensor import Tensor
+
+
+class Embedding(Tensor):
+    ...
