--- docarray/typing/ndarray.py
@@ -1,3 +1,0 @@
-from .tensor import Tensor
-
-Embedding = Tensor
